% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sairr_processes.R
\name{get_sairr_processes}
\alias{get_sairr_processes}
\title{Function to get air quality time series processes serviced by the 
\strong{sairr} package.}
\usage{
get_sairr_processes(file = NA)
}
\arguments{
\item{file}{File of processes helper table.}
}
\value{
Tibble.
}
\description{
Function to get air quality time series processes serviced by the 
\strong{sairr} package.
}
\examples{

# Get processes
data_processes <- get_sairr_processes()

}
\author{
Stuart K. Grange.
}
