% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sairr_simple_summaries.R
\name{get_sairr_simple_summaries}
\alias{get_sairr_simple_summaries}
\title{Function to get simple summaries of air quality observations.}
\usage{
get_sairr_simple_summaries(file = NA, summary = "year", tz = "UTC",
  progress = FALSE)
}
\arguments{
\item{file}{File of simple summaries table.}

\item{summary}{Summary period to import. Can either be \code{"year"} or
\code{"month"}.}

\item{tz}{Time zone for the observations's dates.}

\item{progress}{Should the function give a progress bar for when reading the
data file?}
}
\value{
Tibble.
}
\description{
Function to get simple summaries of air quality observations.
}
\examples{

# Import annual means
data_annual <- get_sairr_simple_summaries(summary = "year")

}
\author{
Stuart K. Grange.
}
