% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sairr_validity.R
\name{get_sairr_summaries}
\alias{get_sairr_summaries}
\title{Function to import summary integers for used in the \strong{sairr} package.}
\usage{
get_sairr_summaries(file = NA)
}
\arguments{
\item{file}{File of summary helper table.}
}
\value{
Tibble.
}
\description{
Function to import summary integers for used in the \strong{sairr} package.
}
\examples{

# Get summary integers
data_validity <- get_sairr_summaries()

}
\author{
Stuart K. Grange.
}
