% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_saq_processes.R
\name{get_saq_processes}
\alias{get_saq_processes}
\title{Function to get air quality time series processes serviced by the 
\strong{saqgetr} package.}
\usage{
get_saq_processes(file = NA)
}
\arguments{
\item{file}{File of processes helper table.}
}
\value{
Tibble.
}
\description{
Function to get air quality time series processes serviced by the 
\strong{saqgetr} package.
}
\examples{

\donttest{

# Get processes
data_processes <- get_saq_processes()

}

}
\author{
Stuart K. Grange.
}
