% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_saq_simple_summaries.R
\name{get_saq_simple_summaries}
\alias{get_saq_simple_summaries}
\title{Function to get simple summaries of air quality observations.}
\usage{
get_saq_simple_summaries(file = NA, summary = "annual_means", tz = "UTC")
}
\arguments{
\item{file}{File of simple summaries table.}

\item{summary}{Summary period to import. Can either be \code{"annual_means"} or
\code{"monthly_means"}.}

\item{tz}{Time zone for the observations' dates.}
}
\value{
Tibble.
}
\description{
Function to get simple summaries of air quality observations.
}
\examples{

\donttest{

# Import annual means
data_annual <- get_saq_simple_summaries(summary = "annual_means")

}

\dontrun{

# Import monthly means, quite a large request so will take some time
data_month <- get_saq_simple_summaries(summary = "monthly_means")

}

}
\author{
Stuart K. Grange.
}
