% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_saq_validity.R
\name{get_saq_summaries}
\alias{get_saq_summaries}
\title{Function to import summary integers for used in the \strong{saqgetr} package.}
\usage{
get_saq_summaries(file = NA)
}
\arguments{
\item{file}{File of summary helper table.}
}
\value{
Tibble.
}
\description{
Function to import summary integers for used in the \strong{saqgetr} package.
}
\examples{

# Get summary integers
data_validity <- get_saq_summaries()

}
\author{
Stuart K. Grange.
}
