% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sairr_observations.R
\name{sairr_observations}
\alias{sairr_observations}
\title{Function to import air quality observations.}
\usage{
sairr_observations(site, tz = "UTC", verbose = FALSE)
}
\arguments{
\item{site}{A vector or sites to import.}

\item{tz}{Time zone for the observations's dates.}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
Function to import air quality observations.
}
\author{
Stuart K. Grange.
}
