% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sairr_processes.R
\name{sairr_processes}
\alias{sairr_processes}
\title{Function to import air quality time series processes serviced by the 
\strong{sairr} package.}
\usage{
sairr_processes(file = NA)
}
\arguments{
\item{file}{File of processes helper table.}
}
\value{
Tibble.
}
\description{
Function to import air quality time series processes serviced by the 
\strong{sairr} package.
}
\author{
Stuart K. Grange.
}
