% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sairr_sites.R
\name{sairr_sites}
\alias{sairr_sites}
\title{Function to import information for monitoring sites/stations/facilities 
serviced by the \strong{sairr} package.}
\usage{
sairr_sites(file = NA)
}
\arguments{
\item{file}{File of sites helper table.}
}
\value{
Tibble.
}
\description{
Function to import information for monitoring sites/stations/facilities 
serviced by the \strong{sairr} package.
}
\examples{

# Load sites table
data_sites <- sairr_sites()

}
\author{
Stuart K. Grange.
}
