% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sairr_validity.R
\name{sairr_summaries}
\alias{sairr_summaries}
\title{Function to import summary integers for used in the \strong{sairr} package.}
\usage{
sairr_summaries(file = NA)
}
\arguments{
\item{file}{File of summary helper table.}
}
\value{
Tibble.
}
\description{
Function to import summary integers for used in the \strong{sairr} package.
}
\author{
Stuart K. Grange.
}
