% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sairr_validity.R
\name{sairr_validity}
\alias{sairr_validity}
\title{Function to import validiy integers for used in the \strong{sairr} package.}
\usage{
sairr_validity(file = NA)
}
\arguments{
\item{file}{File of validiy helper table.}
}
\value{
Tibble.
}
\description{
Function to import validiy integers for used in the \strong{sairr} package.
}
\author{
Stuart K. Grange.
}
