% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz_central.R
\name{tz_central}
\alias{tz_central}
\alias{tz_eastern}
\title{Functions for time-zone strings.}
\usage{
tz_central()

tz_eastern()
}
\description{
Functions for time-zone strings.
}
\author{
Stuart K. Grange
}
