% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sairr_observations.R
\name{get_sairr_observations}
\alias{get_sairr_observations}
\title{Function to get \strong{sairr} air quality observations.}
\usage{
get_sairr_observations(site, tz = "UTC", verbose = FALSE)
}
\arguments{
\item{site}{A vector or sites to import.}

\item{tz}{Time zone for the observations's dates.}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
Function to get \strong{sairr} air quality observations.
}
\examples{

# Load a site's data
data_hafodyrynys <- get_sairr_observations(site = "gb1038a")

}
\author{
Stuart K. Grange.
}
