% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_saq_validity.R
\name{get_saq_validity}
\alias{get_saq_validity}
\title{Function to import validity integers for used in the \strong{saqgetr} package.}
\usage{
get_saq_validity(file = NA)
}
\arguments{
\item{file}{File of validity helper table.}
}
\value{
Tibble.
}
\description{
Function to import validity integers for used in the \strong{saqgetr} package.
}
\examples{

# Get validity integers
data_validity_integers <- get_saq_validity()

}
\author{
Stuart K. Grange.
}
