% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sairr_clean_observations.R
\name{sairr_clean_observations}
\alias{sairr_clean_observations}
\title{Function to clean and format observational data from \strong{sairr}'s 
\code{\link{get_sairr_observations}} function.}
\usage{
sairr_clean_observations(df, summary = "hour", valid_only = TRUE,
  spread = TRUE)
}
\arguments{
\item{df}{Tibble/data frame from \code{\link{get_sairr_observations}}.}

\item{summary}{Summary to filter data to. Default is \code{"hour"}.}

\item{valid_only}{Should only valid observations be kept?}

\item{spread}{Should the data be "spread" where the data frame is reshaped 
so pollutants form variables/columns. This format is usually what is desired
when using \strong{openair}.}
}
\value{
Tibble.
}
\description{
Function to clean and format observational data from \strong{sairr}'s 
\code{\link{get_sairr_observations}} function.
}
\author{
Stuart K. Grange
}
